Adding the GerritHub server for reviews
The Cragsman project was created using the cookiecutter for an
OpenStack project, therefore it got pre-filled with many OpenStack
configurations. However since the project will start first as a PoC
inside the OSIC namespace instead of the openstack namespace, the
gerrit server needs to be updated so the project can have code review.
This commit updates that gerrit server.